requirements.txt
opencv-python==4.5.5.62
numpy==1.19.4
tqdm==4.62.3
h5py==3.1.0
pillow==8.2.0
matplotlib==3.3.4
tensorflow==1.14.0
scipy==1.5.4
pyyaml==5.4.1
six==1.16.0
tensorboard==1.14.0
argparse
easydict==1.9
nibabel==3.2.1
timm==0.4.5 
transforms3d
matplotlib
ninja
open3d==0.9
